gatekeeper-proxy/pyproject.toml
[project]
name = "gatekeeper-proxy"
version = "0.1.0"
description = "Zero-trust reverse proxy with auth, RBAC, and audit logging"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "pydantic-settings>=2.0.0",
    "PyJWT[crypto]>=2.8.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
